+++ CHANGELOG.md
@@ -21,6 +21,7 @@ All notable changes to this project will be documented in this file.
 - *(explorer_card)* Created the card values to return a example of view
 - *(explorer_card)* Created the better card values img to return a example of view
 - *(attachment)* Improved a minimal design on cards
+- *(items)* Load items from localhost api
 
 ### ⚙️ Miscellaneous Tasks